refactor : 테스트 클래스 패키지 이동, Rank 2등인 경우 수정
