+++ src/test/java/lotto/domain/rank/LottoRankTest.java
@@ -1,6 +1,5 @@
-package lotto.domain.lotto;
+package lotto.domain.rank;
 
-import lotto.domain.rank.LottoRank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
