+++ src/test/java/lotto/domain/rank/WinningLottoRankTest.java
@@ -1,9 +1,8 @@
-package lotto.domain.winning;
+package lotto.domain.rank;
 
 import lotto.domain.lotto.LottoTicket;
 import lotto.domain.lotto.LottoTickets;
-import lotto.domain.rank.LottoRank;
-import lotto.domain.rank.WinningLottoRank;
+import lotto.domain.winning.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
